new ip sub-project work on stages new JTransforms

+++ RTlib-stages/src/rtlib/stages/devices/ecc100/variables/PositionVariable.java
@@ -26,7 +26,7 @@ public class PositionVariable extends DoubleVariable
 	@Override
 	public double getEventHook(double pCurrentValue)
 	{
-		int lCurrentPosition = mECC100Axis.getCurrentPosition();
+		double lCurrentPosition = mECC100Axis.getCurrentPosition();
 		return super.getEventHook(lCurrentPosition);
 	}
 }
